Refactor Eeprom to clear Range and not all Eeprom, Move Timer_Index into TimerConstants.h, add LCD into the Window program, add Eeprom BSplinePersistence (NOT TESTED), Improve some EndMatch & StartMatch Device

+++ device/motion/pid/pidDevice.c
@@ -153,7 +153,7 @@ void devicePIDHandleRawData(char commandHeader, InputStream* inputStream, Output
         appendSeparator(outputStream);
 
         // pidType
-        appendHex4(outputStream, computationValues->pidTime);
+        appendHex4(outputStream, (int) computationValues->pidTime);
         appendSeparator(outputStream);
 
         // MotionParameterType
